requirements.txt
Django==4.0.5
psycopg2-binary==2.9.3
gunicorn==20.1.0
Pillow==9.1.1
django-mptt==0.13.4
ipython==8.4.0
python-dateutil==2.8.2
boto3==1.24.28
django-storages==1.12.3
sentry-sdk
xlrd==1.2.0
django-debug-toolbar
